+++ src/logic/tower/types/SimpleTower.h
@@ -5,7 +5,7 @@
 
 class SimpleTower : public Tower {
 public:
-    SimpleTower(WorldState* world_ref, const Position<double>& pos) :
+    SimpleTower(WorldState* world_ref, const Position<int>& pos) :
             Tower(world_ref, 100, 1, 100, 200, 10, pos) {}
 };
